Added some "hooks" to allow greater customization of the new map search. Also addressed a possible security issue. I've been seeing these empty settings, and can only guess that they are attempts by a hacker to do SQL injection. Just to make sure, I now test for admin login before saving any options, even though I'm using the WordPress update_option routine, which should be SQL-injection safe, it is extremely annoying to have dozens of empty settings. If the hacker has elevated to admin, you're screwed, anyway, but this should prevent low-ranked users from causing problems.
git-svn-id: https://plugins.svn.wordpress.org/bmlt-wordpress-satellite-plugin/tags/2.1.29@543565 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ bmlt-wordpress-satellite-plugin.php
@@ -8,13 +8,30 @@
 Plugin Name: BMLT WordPress Satellite
 Plugin URI: http://magshare.org/bmlt
 Description: This is a WordPress plugin satellite of the Basic Meeting List Toolbox.
-Version: 2.1.18
+Version: 2.1.29
 Install: Drop this directory into the "wp-content/plugins/" directory and activate it.
 ********************************************************************************************/
 
-define ( 'BMLT_CURRENT_VERSION', '2.1.18' );    // This needs to be kept in synch with the version above.
+define ( 'BMLT_CURRENT_VERSION', '2.1.29' );    // This needs to be kept in synch with the version above.
 
 // define ( '_DEBUG_MODE_', 1 ); //Uncomment for easier JavaScript debugging.
+
+global $bmlt_localization;  ///< Use this to control the localization.
+
+if ( function_exists ( 'get_option' ) )
+    {
+    $bmlt_localization = get_option ( 'BMLT_lang_base' );
+    }
+
+if ( !isset ( $bmlt_localization ) || !$bmlt_localization )
+    {
+    $bmlt_localization = 'en';  ///< This is the "seed" value for the plugin localization. The first time that you run a localized plugin, change this. It will be reverted to 'en' in the future, but that won't matter anymore.
+    }
+
+if ( function_exists ( 'update_option' ) )
+    {
+    update_option ( 'BMLT_lang_base', $bmlt_localization );
+    }
 
 // Include the satellite driver class.
 require_once ( dirname ( __FILE__ ).'/BMLT-Satellite-Base-Class/bmlt-cms-satellite-plugin.php' );
@@ -194,9 +211,14 @@ class BMLTWPPlugin extends BMLTPlugin
                                         $in_option_value  ///< the values to be set (associative array)
                                         )
         {
-        if ( function_exists ( 'update_option' ) )
+        // Added a test, to prevent the creation of multiple empty settings by low-rank users trying SQL injections.
+        if ( function_exists ( 'update_option' ) && current_user_can ( 'administrator' ) )
             {
             $ret = update_option ( $in_option_key, $in_option_value );
+            }
+        elseif ( !current_user_can ( 'administrator' ) ) 
+            {
+            echo "<!-- BMLTPlugin ERROR (cms_set_option)! Non-admin user attempting to set options! -->";
             }
         else
             {
@@ -210,9 +232,13 @@ class BMLTWPPlugin extends BMLTPlugin
     protected function cms_delete_option ( $in_option_key   ///< The name of the option
                                         )
         {
-        if ( function_exists ( 'delete_option' ) )
+        if ( function_exists ( 'delete_option' ) && current_user_can ( 'administrator' ) )
             {
             $ret = delete_option ( $in_option_key );
+            }
+        elseif ( !current_user_can ( 'administrator' ) ) 
+            {
+            echo "<!-- BMLTPlugin ERROR (cms_set_option)! Non-admin user attempting to delete options! -->";
             }
         else
             {
@@ -327,9 +353,13 @@ class BMLTWPPlugin extends BMLTPlugin
     ****************************************************************************************/
     function option_menu ( )
         {
-        if ( function_exists ( 'add_options_page' ) && (self::get_plugin_object() instanceof BMLTPlugin) )
+        if ( function_exists ( 'add_options_page' ) && (self::get_plugin_object() instanceof BMLTPlugin) && current_user_can ( 'administrator' ) )
             {
             add_options_page ( self::$local_options_title, self::$local_menu_string, 9, basename ( __FILE__ ), array ( self::get_plugin_object(), 'admin_page' ) );
+            }
+        elseif ( !current_user_can ( 'administrator' ) )
+            {
+            echo "<!-- BMLTPlugin ERROR (option_menu)! Non-admin attempting to access the options! -->";
             }
         elseif ( !function_exists ( 'add_options_page' ) )
             {
@@ -385,7 +415,7 @@ class BMLTWPPlugin extends BMLTPlugin
             die ( );
             }
         
-        if ( !$options['gmaps_api_key'] )   // No GMAP API key, no BMLT window.
+        if ( !$options['gmaps_api_key'] || !$this->get_shortcode ( $page->post_content, 'bmlt' ) )   // No GMAP API key, no BMLT window. We also only load the head for the "classic" BMLT.
             {
             $load_head = false;
             }
@@ -506,13 +536,11 @@ class BMLTWPPlugin extends BMLTPlugin
                             )
         {
         // Simple searches can be mixed in with other content.
-        $in_the_content = $this->display_simple_search ( $in_the_content );
+        $in_the_content = parent::content_filter ( $in_the_content );
 
         $count = 0;
 
         $in_the_content = $this->display_popup_search ( $in_the_content, $this->cms_get_post_meta ( get_the_ID(), 'bmlt_simple_searches' ), $count );
-        
-        $in_the_content = $this->display_old_search ( $in_the_content, $count );
         
         return $in_the_content;
         }
@@ -537,9 +565,9 @@ class BMLTWPPlugin extends BMLTPlugin
                 {
                 $changelog = (array) preg_split ( '~[\r\n]+~', trim ( $matches[1] ) );
                 
-                $ret = '<div style="color: #c00;font-size: medium; margin-top:8px;margin-bottom:8px">' . $this->process_text ( $this->plugin_update_message_1 ) . '</div>';
+                $ret = '<div style="color: #c00;font-size: small; margin-top:8px;margin-bottom:8px">' . html_entity_decode ( $this->process_text ( $this->plugin_update_message_1 ) ) . '</div>';
                 $ret .= '<div style="font-weight: normal;">';
-                $ret .= '<p style="margin: 5px 0; font-weight:bold; font-size:medium">' . $this->process_text ( $this->plugin_update_message_2 ) . '</p>';
+                $ret .= '<p style="margin: 5px 0; font-weight:bold; font-size:small">' . html_entity_decode ( $this->process_text ( $this->plugin_update_message_2 ) ) . '</p>';
                 $ul = false;
                 $first = false;
                 
@@ -553,15 +581,15 @@ class BMLTWPPlugin extends BMLTPlugin
                             $ul = true;
                             $first = true;
                             }
-                        $line = preg_replace ('~^\s*\*\s*~', '', $this->process_text ( $line ) );
+                        $line = preg_replace ( '~^\s*\*\s*~', '', $line );
                         if ( $first )
                             {
-                            $ret .= '<li style="list-style-type:none;margin-left: -1.5em; font-weight:bold">' . $this->process_text ( $this->plugin_update_message_3 . ' ' . $line) . '</li>';
+                            $ret .= '<li style="list-style-type:none;margin-left: -1.5em; font-weight:bold">' . html_entity_decode ( $this->process_text ( $this->plugin_update_message_3 ) ) . ' ' . $line . '</li>';
                             $first = false;
                             }
                         else
                             {
-                            $ret .= '<li>' . $this->process_text ( $line ) . '</li>';
+                            $ret .= '<li>' . $line . '</li>';
                             }
                         }
                     else
@@ -571,7 +599,7 @@ class BMLTWPPlugin extends BMLTPlugin
                             $ret .= '</ul><div style="clear: left;"></div>';
                             $ul = false;
                             }
-                        $ret .= '<p style="margin: 5px 0; font-weight:bold; font-size:medium">' . $this->process_text ( $line ) . '</p>';
+                        $ret .= '<p style="margin: 5px 0; font-weight:bold; font-size:small">' . $line . '</p>';
                         }
                     }
